Flytt hentFagsakForPerson fra FagsakContext til ManuellJournalførContext
